refactor(api): split innerworld-region.service.ts into services/innerworld-region/ (api-20q6)
