+++ apps/api/src/routes/innerworld/regions/archive.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../../lib/audit-writer.js";
 import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { archiveRegion } from "../../../services/innerworld-region.service.js";
+import { archiveRegion } from "../../../services/innerworld-region/lifecycle.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
