+++ apps/api/src/routes/innerworld/regions/restore.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { restoreRegion } from "../../../services/innerworld-region.service.js";
+import { restoreRegion } from "../../../services/innerworld-region/lifecycle.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
