+++ apps/api/src/routes/innerworld/regions/update.ts
@@ -7,7 +7,7 @@ import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { updateRegion } from "../../../services/innerworld-region.service.js";
+import { updateRegion } from "../../../services/innerworld-region/update.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
